+++ src/main/kotlin/com/msg/gcms/domain/club/domain/entity/Club.kt
@@ -2,7 +2,7 @@ package com.msg.gcms.domain.club.domain.entity
 
 import com.fasterxml.jackson.annotation.JsonIgnore
 import com.msg.gcms.domain.applicant.domain.entity.Applicant
-import com.msg.gcms.domain.club.enums.ClubType
+import com.msg.gcms.domain.club.domain.entity.enums.ClubType
 import com.msg.gcms.domain.clubMember.domain.entity.ClubMember
 import com.msg.gcms.domain.user.domain.entity.User
 import javax.persistence.*
