+++ src/main/kotlin/com/msg/gcms/domain/club/exception/ClubNotFoundException.kt
@@ -1,4 +1,4 @@
-package com.msg.gcms.domain.club.domain.exception
+package com.msg.gcms.domain.club.exception
 
 import com.msg.gcms.global.exception.ErrorCode
 import com.msg.gcms.global.exception.exceptions.BasicException
