+++ src/main/kotlin/com/msg/gcms/domain/club/presentation/data/dto/ClubIdDto.kt
@@ -1,0 +1,5 @@
+package com.msg.gcms.domain.club.presentation.data.dto
+
+data class ClubIdDto(
+    val clubId: Long
+)
