+++ src/main/kotlin/com/msg/gcms/domain/club/presentation/data/request/ClubIdRequestDto.kt
@@ -1,0 +1,5 @@
+package com.msg.gcms.domain.club.presentation.data.request
+
+data class ClubIdRequestDto(
+    val clubId: Long
+)
