+++ src/main/kotlin/com/msg/gcms/domain/clubMember/domain/entity/ClubMember.kt
@@ -2,7 +2,7 @@ package com.msg.gcms.domain.clubMember.domain.entity
 
 import com.fasterxml.jackson.annotation.JsonIgnore
 import com.msg.gcms.domain.club.domain.entity.Club
-import com.msg.gcms.domain.clubMember.enums.MemberScope
+import com.msg.gcms.domain.clubMember.domain.entity.enums.MemberScope
 import com.msg.gcms.domain.user.domain.entity.User
 import javax.persistence.*
