+++ src/main/kotlin/com/msg/gcms/domain/clubMember/domain/entity/enums/MemberScope.kt
@@ -1,4 +1,4 @@
-package com.msg.gcms.domain.clubMember.enums
+package com.msg.gcms.domain.clubMember.domain.entity.enums
 
 enum class MemberScope(description: String) {
     MEMBER("멤버"),
