--- src/main/kotlin/com/msg/gcms/domain/clubMember/domain/presentation/dto/ClubMemberListDto.kt
@@ -1,9 +1,0 @@
-package com.msg.gcms.domain.clubMember.domain.presentation.dto
-
-import com.msg.gcms.domain.clubMember.enums.MemberScope
-
-class ClubMemberListDto(
-    val scope: MemberScope,
-    val clubMember: List<ClubMemberDto>
-) {
-}
