+++ src/main/kotlin/com/msg/gcms/domain/clubMember/exception/ClubMemberNonExistentException.kt
@@ -1,4 +1,4 @@
-package com.msg.gcms.domain.clubMember.domain.exception
+package com.msg.gcms.domain.clubMember.exception
 
 import com.msg.gcms.global.exception.ErrorCode
 import com.msg.gcms.global.exception.exceptions.BasicException
