+++ src/main/kotlin/com/msg/gcms/domain/clubMember/util/impl/ClubMemberConverterImpl.kt
@@ -1,11 +1,11 @@
-package com.msg.gcms.domain.clubMember.domain.util.impl
+package com.msg.gcms.domain.clubMember.util.impl
 
 import com.msg.gcms.domain.clubMember.domain.entity.ClubMember
-import com.msg.gcms.domain.clubMember.domain.presentation.dto.ClubMemberDto
-import com.msg.gcms.domain.clubMember.domain.presentation.dto.ClubMemberListDto
-import com.msg.gcms.domain.clubMember.domain.presentation.dto.response.ClubMemberListResponse
-import com.msg.gcms.domain.clubMember.domain.util.ClubMemberConverter
-import com.msg.gcms.domain.clubMember.enums.MemberScope
+import com.msg.gcms.domain.clubMember.presentation.data.dto.ClubMemberDto
+import com.msg.gcms.domain.clubMember.presentation.data.dto.ClubMemberListDto
+import com.msg.gcms.domain.clubMember.presentation.data.response.ClubMemberListResponse
+import com.msg.gcms.domain.clubMember.util.ClubMemberConverter
+import com.msg.gcms.domain.clubMember.domain.entity.enums.MemberScope
 import org.springframework.stereotype.Component
 
 @Component
